pipeline/requirements.txt
mssql-python
python-dotenv
pylint
pytest
mssql-python
python-dotenv
pandas
requests
